# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoraPay API",
    "version": "0.1.0",
    "description": "Cross-border remittance platform. Russia/CIS → Nigeria & Ghana.\n\n**No live funds.** `LIVE_FUNDS_ENABLED` defaults to false and every provider\nruns against a simulator until partner agreements and written legal opinions\nare in place.\n\nTwo conventions that differ from most payment APIs, and both are deliberate:\n\n1. **Money is never a JSON number.** Amounts carry `minorUnits` as a string,\n   which is the authoritative field. A JSON number is a double, and a double\n   is not a currency.\n2. **An acknowledgement is not a settlement.** A provider that accepts a\n   submission has moved no money. Only a status poll or the T+1 statement\n   produces an outcome, and only an outcome posts to the ledger.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.staging.morapay.example",
      "description": "Staging"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Back-office tokens are signed with a different key from customer tokens. A customer token fails signature verification here, not merely authorisation."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/transfers/{id}": {
      "get": {
        "tags": [
          "Transfers"
        ],
        "summary": "Transfer status and timeline",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "QUOTED",
                        "COMPLIANCE_PENDING",
                        "ON_HOLD",
                        "AWAITING_PAYIN",
                        "PAYIN_CONFIRMED",
                        "SETTLING",
                        "PAYOUT_INITIATED",
                        "PAYOUT_CONFIRMED",
                        "COMPLETED",
                        "REFUNDING",
                        "REFUNDED",
                        "FAILED"
                      ]
                    },
                    "senderStatus": {
                      "type": "string"
                    },
                    "corridorId": {
                      "type": "string"
                    },
                    "purpose": {
                      "type": "string",
                      "enum": [
                        "FAMILY_SUPPORT",
                        "EDUCATION",
                        "MEDICAL",
                        "GIFT",
                        "OWN_ACCOUNT"
                      ]
                    },
                    "sendAmount": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "string"
                        },
                        "currency": {
                          "type": "string"
                        },
                        "minorUnits": {
                          "type": "string"
                        },
                        "formatted": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "amount",
                        "currency",
                        "minorUnits",
                        "formatted"
                      ]
                    },
                    "fee": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "string"
                        },
                        "currency": {
                          "type": "string"
                        },
                        "minorUnits": {
                          "type": "string"
                        },
                        "formatted": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "amount",
                        "currency",
                        "minorUnits",
                        "formatted"
                      ]
                    },
                    "totalToPay": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "string"
                        },
                        "currency": {
                          "type": "string"
                        },
                        "minorUnits": {
                          "type": "string"
                        },
                        "formatted": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "amount",
                        "currency",
                        "minorUnits",
                        "formatted"
                      ]
                    },
                    "recipientAmount": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "string"
                        },
                        "currency": {
                          "type": "string"
                        },
                        "minorUnits": {
                          "type": "string"
                        },
                        "formatted": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "amount",
                        "currency",
                        "minorUnits",
                        "formatted"
                      ]
                    },
                    "recipient": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "maskedAccount": {
                          "type": "string"
                        },
                        "resolvedName": {
                          "type": "string",
                          "nullable": true
                        },
                        "country": {
                          "type": "string"
                        },
                        "method": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "maskedAccount",
                        "resolvedName",
                        "country",
                        "method"
                      ]
                    },
                    "payinMethod": {
                      "type": "string",
                      "enum": [
                        "SBP",
                        "QR",
                        "CARD",
                        "VIRTUAL_ACCOUNT"
                      ]
                    },
                    "payinInstructions": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": {
                            "kind": {
                              "type": "string",
                              "enum": [
                                "SBP"
                              ]
                            },
                            "deepLink": {
                              "type": "string"
                            },
                            "expiresAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "kind",
                            "deepLink",
                            "expiresAt"
                          ]
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": {
                              "type": "string",
                              "enum": [
                                "QR"
                              ]
                            },
                            "payload": {
                              "type": "string"
                            },
                            "expiresAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "kind",
                            "payload",
                            "expiresAt"
                          ]
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": {
                              "type": "string",
                              "enum": [
                                "VIRTUAL_ACCOUNT"
                              ]
                            },
                            "accountNumber": {
                              "type": "string"
                            },
                            "bankName": {
                              "type": "string"
                            },
                            "reference": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "kind",
                            "accountNumber",
                            "bankName",
                            "reference"
                          ]
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": {
                              "type": "string",
                              "enum": [
                                "CARD"
                              ]
                            },
                            "redirectUrl": {
                              "type": "string"
                            },
                            "expiresAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "kind",
                            "redirectUrl",
                            "expiresAt"
                          ]
                        },
                        {
                          "nullable": true
                        }
                      ]
                    },
                    "failureCode": {
                      "type": "string",
                      "nullable": true
                    },
                    "failureReason": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "completedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "timeline": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "event": {
                            "type": "string"
                          },
                          "fromState": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "QUOTED",
                              "COMPLIANCE_PENDING",
                              "ON_HOLD",
                              "AWAITING_PAYIN",
                              "PAYIN_CONFIRMED",
                              "SETTLING",
                              "PAYOUT_INITIATED",
                              "PAYOUT_CONFIRMED",
                              "COMPLETED",
                              "REFUNDING",
                              "REFUNDED",
                              "FAILED"
                            ]
                          },
                          "toState": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "QUOTED",
                              "COMPLIANCE_PENDING",
                              "ON_HOLD",
                              "AWAITING_PAYIN",
                              "PAYIN_CONFIRMED",
                              "SETTLING",
                              "PAYOUT_INITIATED",
                              "PAYOUT_CONFIRMED",
                              "COMPLETED",
                              "REFUNDING",
                              "REFUNDED",
                              "FAILED"
                            ]
                          },
                          "senderStatus": {
                            "type": "string"
                          },
                          "at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "event",
                          "fromState",
                          "toState",
                          "senderStatus",
                          "at"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "reference",
                    "state",
                    "senderStatus",
                    "corridorId",
                    "purpose",
                    "sendAmount",
                    "fee",
                    "totalToPay",
                    "recipientAmount",
                    "recipient",
                    "payinMethod",
                    "payinInstructions",
                    "failureCode",
                    "failureReason",
                    "createdAt",
                    "updatedAt",
                    "completedAt",
                    "timeline"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}